backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
